MyPortfolio/portfolio_project/requirements.txt
# Core Django
Django==4.2.7
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# AI & API
google-generativeai==0.5.4

# Media & Files
Pillow==10.0.0
django-storages==1.14.2
boto3==1.34.0

# Web Server & Production
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Security & CORS
django-cors-headers==4.3.1
cryptography==41.0.7

# Environment & Config
python-decouple==3.8

# Monitoring & Logging
sentry-sdk==1.38.0

# Code Quality (Optional but recommended)
black==23.12.0
flake8==6.1.0
django-debug-toolbar==4.2.0

# Testing (Optional)
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0